Add new option `--png` to generate a bitmap screenshot in favor of a PDF
This allows using puppeteer-pdf as a replacement for wkhtmltoimage in addition
to wkhtmltopdf. This is not equivalent to converting first to PDF and then to
PNG because the image is generated to cover the entire page without any page
breaks.

Example invocation:

	CHROME_BIN=/usr/bin/chromium ./puppeteer-pdf.js --png -w 800 -p test.png https://masysma.net/31/web_main.xhtml

So far, the only supported options together with `--png` are the following ones:
-w, -h, -oBg, -p and -s